requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
pymysql
cryptography
pydantic
passlib[bcrypt]
pyjwt
